# docs/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "description": "Yagona xato formati (BE-1801). Qo'shimcha maydonlar koddan kelib chiqadi.",
        "properties": {
          "attempts_left": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempts Left"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "retry_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "OnboardingOut": {
        "properties": {
          "balance_set": {
            "title": "Balance Set",
            "type": "boolean"
          },
          "has_goals": {
            "title": "Has Goals",
            "type": "boolean"
          },
          "has_reminders": {
            "title": "Has Reminders",
            "type": "boolean"
          },
          "has_transactions": {
            "title": "Has Transactions",
            "type": "boolean"
          }
        },
        "required": [
          "balance_set",
          "has_transactions",
          "has_goals",
          "has_reminders"
        ],
        "title": "OnboardingOut",
        "type": "object"
      },
      "OtpPurpose": {
        "enum": [
          "login",
          "pin_reset"
        ],
        "title": "OtpPurpose",
        "type": "string"
      },
      "OtpRequestIn": {
        "additionalProperties": false,
        "properties": {
          "attestation_token": {
            "anyOf": [
              {
                "maxLength": 8192,
                "minLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attestation Token"
          },
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "phone": {
            "maxLength": 20,
            "minLength": 9,
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "phone",
          "device_id"
        ],
        "title": "OtpRequestIn",
        "type": "object"
      },
      "OtpSentOut": {
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "resend_after": {
            "title": "Resend After",
            "type": "integer"
          }
        },
        "required": [
          "resend_after",
          "expires_in"
        ],
        "title": "OtpSentOut",
        "type": "object"
      },
      "OtpVerifyIn": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^\\d{6}$",
            "title": "Code",
            "type": "string"
          },
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "device_name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Device Name",
            "type": "string"
          },
          "device_public_key": {
            "maxLength": 700,
            "minLength": 40,
            "pattern": "^[A-Za-z0-9_\\-+/=]+$",
            "title": "Device Public Key",
            "type": "string"
          },
          "phone": {
            "maxLength": 20,
            "minLength": 9,
            "title": "Phone",
            "type": "string"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "purpose": {
            "$ref": "#/components/schemas/OtpPurpose",
            "default": "login"
          }
        },
        "required": [
          "phone",
          "code",
          "device_id",
          "device_public_key",
          "device_name",
          "platform"
        ],
        "title": "OtpVerifyIn",
        "type": "object"
      },
...
      "Platform": {
        "enum": [
          "ios",
          "android"
        ],
        "title": "Platform",
        "type": "string"
      },
...
      "RefreshIn": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "maxLength": 128,
            "minLength": 20,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshIn",
        "type": "object"
      },
...
      "TokenOut": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenOut",
        "type": "object"
      },
...
      "VerifyOut": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/VerifyUserOut"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "user"
        ],
        "title": "VerifyOut",
        "type": "object"
      },
      "VerifyUserOut": {
        "properties": {
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "has_pin_setup": {
            "title": "Has Pin Setup",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_new": {
            "title": "Is New",
            "type": "boolean"
          },
          "onboarding": {
            "$ref": "#/components/schemas/OnboardingOut"
          }
        },
        "required": [
          "id",
          "first_name",
          "is_new",
          "has_pin_setup",
          "onboarding"
        ],
        "title": "VerifyUserOut",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Finora API v1. Barcha vaqtlar ISO 8601 UTC; mijoz vaqt zonasi `X-Timezone` sarlavhasida. Summalar — so'm (butun son). Xato formati: `{code, message, request_id}` (+ `fields`, `retry_after`, `attempts_left`). Yozuv yaratuvchi POST'lar `Idempotency-Key` talab qiladi. Xato kodlari: docs/backend/ERROR_CODES.md.",
    "title": "Finora API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout_v1_auth_logout_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/otp": {
      "post": {
        "description": "Raqam ro'yxatdan o'tgan-o'tmaganidan qat'i nazar bir xil javob (BE-101).",
        "operationId": "request_otp_v1_auth_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpSentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "Request Otp",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "description": "Qurilma kaliti bilan imzolangan (X-Device-Timestamp, X-Device-Signature).",
        "operationId": "refresh_v1_auth_refresh_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-device-timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "x-device-signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Signature"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/verify": {
      "post": {
        "operationId": "verify_otp_v1_auth_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "Verify Otp",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}